# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Arxivist Backend API
  version: 1.0.0
  description: |
    Бэкенд-модуль HeadHunter-агента (HeadHunter + интернет + базы знаний) с доступом по API-ключу.

    ## Аутентификация
    Все эндпоинты (кроме `/docs`, `/openapi.yaml`, `/health`) требуют API-ключ в заголовке
    `X-API-Key` (или `Authorization: Bearer <key>`). Ключи выпускаются оператором:
    `python scripts/apikeys.py create --label "<имя клиента>"`. Ключ показывается один раз;
    в хранилище лежит только его SHA-256. revoke действует немедленно, без рестарта.

    ## Два семейства эндпоинтов
    1. **Агентские задачи** (`/agent/jobs`) — полный LLM-агент со всеми интеграциями
       (HeadHunter, базы знаний, arXiv, веб-поиск, дашборды). Выполняются минуты, поэтому
       асинхронно: `POST` возвращает `202 {job_id}`, дальше поллинг `GET /agent/jobs/{id}`
       (рекомендуемый интервал 2–5 с). Результат — markdown-ответ агента.
    2. **Типизированные эндпоинты** (`/hh/*`, `/kb/*`, `/candidates/save`, `/arxiv/search`) —
       детерминированные вызовы без LLM, отвечают синхронно за секунды
       (исключение — `/arxiv/search` и `/kb/query`, до 1–3 минут).

    ## Коды ошибок
    Все не-2xx ответы имеют форму `{"error": {"code", "message", ...}}`:
    | HTTP | code | смысл |
    |---|---|---|
    | 400 | `invalid_request` | неверные параметры запроса |
    | 401 | `invalid_api_key` | ключ отсутствует/неверен/отозван |
...
    | 404 | `job_not_found` / `file_not_found` | нет такого объекта |
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /agent/jobs:
    post:
      tags: [agent]
      summary: Создать агентскую задачу
      description: |
        Запускает полный агентский флоу — тот же, что в чате: подбор кандидатов по
        текстовым критериям или по приложенному документу вакансии (`attached_file_ids`
        из `POST /files`), обогащение базы знаний, генерация дашбордов, исследования.
        Ответ приходит через поллинг `GET /agent/jobs/{job_id}`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentJobRequest" }
            examples:
              candidates:
                summary: Подбор кандидатов по критериям
                value:
                  message: "Найди Python-разработчиков в Москве с опытом LLM, покажи топ-10 с ссылками на резюме"
              vacancy_doc:
                summary: Подбор по документу вакансии
                value:
                  message: "Подбери кандидатов под прикреплённое описание вакансии"
                  attached_file_ids: ["0cd042478f654210a1f8d83491f8a651"]
              followup:
                summary: Продолжение диалога (история передаётся клиентом)
                value:
                  message: "Сохрани кандидатов 1 и 3"
                  session_id: "my-session-42"
                  history:
                    - { role: "user", content: "Найди Python-разработчиков в Москве" }
                    - { role: "assistant", content: "Нашлось 138 резюме... 1. ... 2. ..." }
      responses:
        "202":
          description: Задача принята
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentJobAccepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429":
          description: Ёмкость занята — повторите позже (заголовок Retry-After)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /agent/jobs/{job_id}:
    get:
      tags: [agent]
      summary: Статус/результат задачи
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: |
            Текущее состояние. `state`: `queued` → `running` → `succeeded | failed | canceled`.
            При `running` поле `progress` содержит человекочитаемые фразы прогресса.
            При `succeeded` поле `result.reply` — итоговый markdown-ответ агента.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentJobStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [agent]
      summary: Отменить задачу
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Задача отменена (или уже была завершена)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentJobStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /files:
    post:
      tags: [files]
      summary: Загрузить документ (PDF/DOCX/MD/TXT, ≤25 МБ)
      description: Полученные `file_id` передаются в `attached_file_ids` агентской задачи.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Метаданные сохранённых файлов
          content:
            application/json:
              example:
                files:
                  - { file_id: "0cd0424…", name: "vacancy.md", parse_mode: plain, char_count: 328, size: 508 }
                errors: []
        "401": { $ref: "#/components/responses/Unauthorized" }
  /files/{name}:
    get:
      tags: [files]
      summary: Скачать полученный PDF (например, статью arXiv)
      parameters:
        - { name: name, in: path, required: true, schema: { type: string }, example: "2401.12345v1.pdf" }
      responses:
        "200":
          description: PDF-файл
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /hh/vacancies:
    get:
      tags: [headhunter]
      summary: Поиск вакансий
      description: Возвращает сырой JSON api.hh.ru (`items`, `found`, `pages`, `page`).
      parameters:
        - { name: text, in: query, schema: { type: string }, example: "python developer" }
        - { name: area, in: query, schema: { type: integer }, description: "1=Москва, 2=СПб, 113=Россия" }
        - { name: employer_id, in: query, schema: { type: integer } }
        - { name: experience, in: query, schema: { type: string, enum: [noExperience, between1And3, between3And6, moreThan6] } }
        - { name: employment, in: query, schema: { type: string, enum: [full, part, project, volunteer, probation] } }
        - { name: schedule, in: query, schema: { type: string, enum: [fullDay, shift, flexible, remote, flyInFlyOut] } }
        - { name: salary, in: query, schema: { type: integer }, description: "минимальная зарплата, RUB" }
        - { name: order_by, in: query, schema: { type: string, enum: [relevance, publication_time, salary_desc, salary_asc] } }
        - { name: per_page, in: query, schema: { type: integer, maximum: 100, default: 20 } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: date_from, in: query, schema: { type: string }, example: "2026-07-01T00:00:00+0300" }
      responses:
        "200": { $ref: "#/components/responses/HHRaw" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /hh/vacancies/{vacancy_id}:
    get:
      tags: [headhunter]
      summary: Полная вакансия по ID
      parameters:
        - { name: vacancy_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/HHRaw" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /hh/employers:
    get:
      tags: [headhunter]
      summary: Поиск компаний
      parameters:
        - { name: text, in: query, schema: { type: string }, example: "Яндекс" }
        - { name: area, in: query, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string }, example: "company" }
        - { name: per_page, in: query, schema: { type: integer, maximum: 100, default: 20 } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200": { $ref: "#/components/responses/HHRaw" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /hh/employers/{employer_id}:
    get:
      tags: [headhunter]
      summary: Профиль компании (+ опционально вакансии)
      parameters:
        - { name: employer_id, in: path, required: true, schema: { type: string } }
        - { name: include_vacancies, in: query, schema: { type: boolean, default: false },
            description: "true — добавить в ответ до 20 открытых вакансий (поле vacancies)" }
      responses:
        "200": { $ref: "#/components/responses/HHRaw" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
...
  /hh/resumes/{resume_id}:
    get:
      tags: [headhunter]
      summary: Полное резюме по ID
      parameters:
        - { name: resume_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/HHRaw" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/HHUserAuthRequired" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /hh/reference/{ref_type}:
    get:
      tags: [headhunter]
      summary: Справочники (регионы, роли, словари, навыки)
      parameters:
        - { name: ref_type, in: path, required: true,
            schema: { type: string, enum: [areas, professional_roles, dictionaries, skills] } }
        - { name: query, in: query, schema: { type: string },
            description: "обязателен для ref_type=skills (автодополнение навыка)" }
      responses:
        "200": { $ref: "#/components/responses/HHRaw" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /hh/auth/status:
    get:
      tags: [headhunter]
      summary: Статус авторизации HeadHunter
      description: Показывает, есть ли пользовательский токен для поиска резюме.
      responses:
        "200":
          description: Статус токена
          content:
            application/json:
              example:
                configured: true
                grant: authorization_code
                has_refresh_token: true
                expires_in_seconds: 1200000
                user_authorized: true
                authorize_url: "/arxivist/hh/authorize"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /kb/databases:
    get:
      tags: [knowledge-base]
      summary: Список баз знаний
      responses:
        "200":
          description: Базы
          content:
            application/json:
              example: { databases: [ { name: companies, status: ready }, { name: candidates, status: ready } ] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
...
  /kb/tasks/{task_id}:
    get:
      tags: [knowledge-base]
      summary: Статус фоновой задачи базы знаний
      parameters:
        - { name: task_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Состояние (SUCCESS / FAILURE / PROGRESS)
          content:
            application/json:
              example: { task_id: "abc123", state: SUCCESS, progress: 100, result: { files_processed: 1 }, error: null }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "Ключ вида axv_<id>_<secret>, выпущенный scripts/apikeys.py"
    BearerAuth:
      type: http
      scheme: bearer
      description: "Тот же ключ в Authorization: Bearer <key>"
...